# Conflicts:
#	Insert_QLNhaSach1.sql
#	QuanLyHeThongNhaSach/system/BUS/.vs/BUS.csproj.dtbcache.json
#	QuanLyHeThongNhaSach/system/BUS/BUS.csproj
#	QuanLyHeThongNhaSach/system/BUS/BUS_HangHoa.cs
#	QuanLyHeThongNhaSach/system/BUS/bin/Debug/BUS.dll
#	QuanLyHeThongNhaSach/system/BUS/bin/Debug/BUS.pdb
#	QuanLyHeThongNhaSach/system/BUS/bin/Debug/DAL.dll
#	QuanLyHeThongNhaSach/system/BUS/bin/Debug/DAL.pdb
#	QuanLyHeThongNhaSach/system/BUS/bin/Debug/ET.dll
#	QuanLyHeThongNhaSach/system/BUS/bin/Debug/ET.pdb
#	QuanLyHeThongNhaSach/system/BUS/obj/Debug/BUS.csproj.AssemblyReference.cache
#	QuanLyHeThongNhaSach/system/BUS/obj/Debug/BUS.csproj.CoreCompileInputs.cache
#	QuanLyHeThongNhaSach/system/BUS/obj/Debug/BUS.csproj.FileListAbsolute.txt
#	QuanLyHeThongNhaSach/system/BUS/obj/Debug/BUS.dll
#	QuanLyHeThongNhaSach/system/BUS/obj/Debug/BUS.pdb
#	QuanLyHeThongNhaSach/system/BUS/obj/Debug/DesignTimeResolveAssemblyReferencesInput.cache
#	QuanLyHeThongNhaSach/system/DAL/.vs/DAL.csproj.dtbcache.json
#	QuanLyHeThongNhaSach/system/DAL/DAL.csproj
#	QuanLyHeThongNhaSach/system/DAL/DAL_Data.cs
#	QuanLyHeThongNhaSach/system/DAL/DAL_HangHoa.cs
#	QuanLyHeThongNhaSach/system/DAL/QLNhaSach.dbml
#	QuanLyHeThongNhaSach/system/DAL/QLNhaSach.dbml.layout
#	QuanLyHeThongNhaSach/system/DAL/QLNhaSach.designer.cs
#	QuanLyHeThongNhaSach/system/DAL/bin/Debug/DAL.dll
#	QuanLyHeThongNhaSach/system/DAL/bin/Debug/DAL.pdb
#	QuanLyHeThongNhaSach/system/DAL/bin/Debug/ET.dll
#	QuanLyHeThongNhaSach/system/DAL/bin/Debug/ET.pdb
#	QuanLyHeThongNhaSach/system/DAL/obj/Debug/DAL.csproj.AssemblyReference.cache
#	QuanLyHeThongNhaSach/system/DAL/obj/Debug/DAL.csproj.CoreCompileInputs.cache
#	QuanLyHeThongNhaSach/system/DAL/obj/Debug/DAL.csproj.FileListAbsolute.txt
#	QuanLyHeThongNhaSach/system/DAL/obj/Debug/DAL.dll
#	QuanLyHeThongNhaSach/system/DAL/obj/Debug/DAL.pdb
#	QuanLyHeThongNhaSach/system/DAL/obj/Debug/DesignTimeResolveAssemblyReferencesInput.cache
#	QuanLyHeThongNhaSach/system/DAL/obj/Debug/TempPE/QLNhaSach.designer.cs.dll
#	QuanLyHeThongNhaSach/system/ET/.vs/ET.csproj.dtbcache.json
#	QuanLyHeThongNhaSach/system/ET/ET.csproj
#	QuanLyHeThongNhaSach/system/ET/bin/Debug/ET.dll
#	QuanLyHeThongNhaSach/system/ET/bin/Debug/ET.pdb
#	QuanLyHeThongNhaSach/system/ET/obj/Debug/DesignTimeResolveAssemblyReferencesInput.cache
#	QuanLyHeThongNhaSach/system/ET/obj/Debug/ET.csproj.CoreCompileInputs.cache
#	QuanLyHeThongNhaSach/system/ET/obj/Debug/ET.csproj.FileListAbsolute.txt
#	QuanLyHeThongNhaSach/system/ET/obj/Debug/ET.dll
#	QuanLyHeThongNhaSach/system/ET/obj/Debug/ET.pdb
#	QuanLyHeThongNhaSach/system/GUI/.vs/GUI.csproj.dtbcache.json
#	QuanLyHeThongNhaSach/system/GUI/.vs/GUI/v17/.suo
#	QuanLyHeThongNhaSach/system/GUI/.vs/GUI/v17/DocumentLayout.backup.json
#	QuanLyHeThongNhaSach/system/GUI/.vs/GUI/v17/DocumentLayout.json
#	QuanLyHeThongNhaSach/system/GUI/QuanLyCaTruc.Designer.cs
#	QuanLyHeThongNhaSach/system/GUI/bin/Debug/BUS.dll
#	QuanLyHeThongNhaSach/system/GUI/bin/Debug/BUS.pdb
#	QuanLyHeThongNhaSach/system/GUI/bin/Debug/DAL.dll
#	QuanLyHeThongNhaSach/system/GUI/bin/Debug/DAL.pdb
#	QuanLyHeThongNhaSach/system/GUI/bin/Debug/ET.dll
#	QuanLyHeThongNhaSach/system/GUI/bin/Debug/ET.pdb
#	QuanLyHeThongNhaSach/system/GUI/bin/Debug/GUI.exe
#	QuanLyHeThongNhaSach/system/GUI/bin/Debug/GUI.pdb
#	QuanLyHeThongNhaSach/system/GUI/formThanhToan.Designer.cs
#	QuanLyHeThongNhaSach/system/GUI/obj/Debug/DesignTimeResolveAssemblyReferencesInput.cache
#	QuanLyHeThongNhaSach/system/GUI/obj/Debug/GUI.csproj.AssemblyReference.cache
#	QuanLyHeThongNhaSach/system/GUI/obj/Debug/GUI.csproj.CoreCompileInputs.cache
#	QuanLyHeThongNhaSach/system/GUI/obj/Debug/GUI.csproj.FileListAbsolute.txt
#	QuanLyHeThongNhaSach/system/GUI/obj/Debug/GUI.csproj.GenerateResource.cache
#	QuanLyHeThongNhaSach/system/GUI/obj/Debug/GUI.exe
#	QuanLyHeThongNhaSach/system/GUI/obj/Debug/GUI.pdb

+++ QLNS(1.1).sql
@@ -58,7 +58,7 @@ create table HangHoa (
 maHH varchar(13) not null,
 maLHH varchar(5) foreign key references LoaiHangHoa(maLHH) ,
 tenHH nvarchar(45) not null,
-giaHH int not null,
+giaHH float not null,
 donViTinh nvarchar(15) not null,
 soLuongTon int not null,
 maNPP varchar(11) foreign key references NhaPhanPhoi(maNPP),
